Fixes MW's Calibration Bug plus several minor issues
    - Changes uint16_t* pointers/memory throughout to be double* (fixing calibration bug)
    - Removes thresholds from histogram functions (never needed)
    - Removes unused pointers/variables
    - Replaces calls to delete() were the memory had been calloc'd (always use free())
    - Fixes HxtProcessingLibTest not picking up Pixel_Grid from .ini configuration file
    - Fixes HxtProcessingLibTest crashing if Output_Prefix in .ini file contained either '\' or '/'
    - Fixes data acquisition tab to now update all of its' GUI items upon start-up

+++ HxtProcessingLib/generalframeprocessor.h
@@ -17,18 +17,18 @@ public:
    void setEnergyCalibration(bool energyCalibration);
    GeneralPixelProcessor *getPixelProcessor();
 
-   virtual uint16_t *process(unordered_map<int, double>**pixelRawValMapPtr, 
+   virtual double *process(unordered_map<int, double>**pixelRawValMapPtr,
                              uint16_t *frame) = 0;
-   virtual uint16_t *process(unordered_map<int, double>**pixelRawValMapPtr, 
+   virtual double *process(unordered_map<int, double>**pixelRawValMapPtr,
                              uint16_t *frame, uint16_t thresholdValue) = 0;
-   virtual uint16_t *process(unordered_map<int, double>**pixelRawValMapPtr, 
+   virtual double *process(unordered_map<int, double>**pixelRawValMapPtr,
                              uint16_t *frame, uint16_t *thresholdPerPixel) = 0;
 
-   virtual uint16_t *process(uint16_t *frame,
+   virtual double *process(uint16_t *frame,
                              unordered_map<int, double>**pixelEnergyMapPtr) = 0;
-   virtual uint16_t *process(uint16_t *frame, uint16_t thresholdValue,
+   virtual double *process(uint16_t *frame, uint16_t thresholdValue,
                              unordered_map<int, double>**pixelEnergyMapPtr) = 0;
-   virtual uint16_t *process(uint16_t *frame, uint16_t *thresholdPerPixel,
+   virtual double *process(uint16_t *frame, uint16_t *thresholdPerPixel,
                              unordered_map<int, double>**pixelEnergyMapPtr) = 0;
 
 protected:
