+++ HxtProcessingLib/generalpixelprocessor.h
@@ -30,29 +30,29 @@ public:
    virtual void resetLastRe_orderedSize();
    ///
 
-   uint16_t *processFrame(unordered_map<int, double>**pixelRawValMapPtr,
+   double *processFrame(unordered_map<int, double>**pixelRawValMapPtr,
                           uint16_t *frame);
-   virtual uint16_t *processFrame(unordered_map<int, double>**pixelRawValMapPtr,
+   virtual double *processFrame(unordered_map<int, double>**pixelRawValMapPtr,
                                   uint16_t *frame, uint16_t thresholdValue);
-   virtual uint16_t *processFrame(unordered_map<int, double>**pixelRawValMapPtr,
+   virtual double *processFrame(unordered_map<int, double>**pixelRawValMapPtr,
                                   uint16_t *frame, uint16_t *thresholdPerPixel);
 
-   virtual uint16_t *processFrame(uint16_t *frame,
+   virtual double *processFrame(uint16_t *frame,
               unordered_map<int, double>**pixelEnergyMapPtr);
-   virtual uint16_t *processFrame(uint16_t *frame, uint16_t thresholdValue,
+   virtual double *processFrame(uint16_t *frame, uint16_t thresholdValue,
 				  unordered_map<int, double>**pixelEnergyMapPtr);
-   virtual uint16_t *processFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
+   virtual double *processFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
 				  unordered_map<int, double>**pixelEnergyMapPtr);
 
-   uint16_t *processRe_orderFrame(unordered_map<int, double>**pixelRawValMapPtr, uint16_t *frame);
-   virtual uint16_t *processRe_orderFrame(unordered_map<int, double>**pixelRawValMapPtr, uint16_t *frame, uint16_t thresholdValue);
-   virtual uint16_t *processRe_orderFrame(unordered_map<int, double>**pixelRawValMapPtr, uint16_t *frame, uint16_t *thresholdPerPixel);
+   double *processRe_orderFrame(unordered_map<int, double>**pixelRawValMapPtr, uint16_t *frame);
+   virtual double *processRe_orderFrame(unordered_map<int, double>**pixelRawValMapPtr, uint16_t *frame, uint16_t thresholdValue);
+   virtual double *processRe_orderFrame(unordered_map<int, double>**pixelRawValMapPtr, uint16_t *frame, uint16_t *thresholdPerPixel);
 
-   uint16_t *processRe_orderFrame(uint16_t *frame, 
+   double *processRe_orderFrame(uint16_t *frame,
 				  unordered_map<int, double>**pixelEnergyMapPtr);
-   virtual uint16_t *processRe_orderFrame(uint16_t *frame, uint16_t thresholdValue,
+   virtual double *processRe_orderFrame(uint16_t *frame, uint16_t thresholdValue,
                                                   unordered_map<int, double>**pixelEnergyMapPtr);
-   virtual uint16_t *processRe_orderFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
+   virtual double *processRe_orderFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
               unordered_map<int, double>**pixelEnergyMapPtr);
 
 protected:
