+++ HxtProcessingLib/pixelprocessornextframe.h
@@ -16,21 +16,21 @@ public:
 
    void resetLastRe_orderedSize();
 
-   uint16_t *processFrame(uint16_t *frame, uint16_t thresholdValue);
-   uint16_t *processFrame(uint16_t *frame, uint16_t *thresholdPerPixel);
+   double *processFrame(uint16_t *frame, uint16_t thresholdValue);
+   double *processFrame(uint16_t *frame, uint16_t *thresholdPerPixel);
 
-   uint16_t *processFrame(uint16_t *frame, uint16_t thresholdValue,
+   double *processFrame(uint16_t *frame, uint16_t thresholdValue,
                                   unordered_map<int, double>**pixelEnergyMapPtr);
-   uint16_t *processFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
+   double *processFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
                                   unordered_map<int, double>**pixelEnergyMapPtr);
 
   
-   uint16_t *processRe_orderFrame(uint16_t *frame, uint16_t thresholdValue);
-   uint16_t *processRe_orderFrame(uint16_t *frame, uint16_t *thresholdPerPixel);
+   double *processRe_orderFrame(uint16_t *frame, uint16_t thresholdValue);
+   double *processRe_orderFrame(uint16_t *frame, uint16_t *thresholdPerPixel);
 
-   virtual uint16_t *processRe_orderFrame(uint16_t *frame, uint16_t thresholdValue,
+   virtual double *processRe_orderFrame(uint16_t *frame, uint16_t thresholdValue,
                                   unordered_map<int, double>**pixelEnergyMapPtr);
-   uint16_t *processRe_orderFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
+   double *processRe_orderFrame(uint16_t *frame, uint16_t *thresholdPerPixel,
                                   unordered_map<int, double>**pixelEnergyMapPtr);
 
 private:
